✨ src: Implemented <AppDatepickerDialog>
Finally got some time to implement yet another important piece of the
element which is the dialog version. It has been a tough lesson for me
to get some stuffs to work as expected. `Web Animations` is used in this
case for animations basically but the weird behavior from both the
native supported browsers and the polyfill created different end
results. Thanks to a simple polyfill for browsers like Chrome lacks
native supported for the `Promise`d version that exposes a property
named `finished`. By loading polyfill just because of that, it breaks
all the animations on Chrome but it worked fine on other browsers. BTW,
Firefox is the only browser that shipped a working Web Animations on its
stable version.

Another new update is that `rollup` might be redundant for creating
libraries. `TypeScript`'s own CLI suffice to do the compilation since
bundling is not needed for `lit-element` based components.