💄 src: Minor CSS update to match that of native Android datepicker
Minor CSS update including `border-radius` and improved animations when
opening/ closing dialog.

Removed `finished` polyfill for Chrome or browsers that natively
supported Web Animations. A simple solution for that is to wrap the
`onfinish` event handler with a Promise like so:

  new Promise(yay => (fadeInAnimation.onfinish = yay)).then(...)

This has been a technique used while developing the datepicker. Totally
forgotten about that and now this one-liner solution is once back again
for the dialog's animations.

Also worth to document this down later. `web-animations-next-lite` is
the recommended polyfill for better cross platform compatibility. It
works perfectly on Edge, Firefox, and Chrome (all tested). Yet to do a
testing on Safari. It's TP version supports native Web Animations behind
experimental flag. The main issue in other polyfill version is that
`fill` is causing discrepancies after applying polyfills. On Edge,
animated element will be set to the previous state forever and never get
reset for subsequent animation to take place. In the case of datepicker,
tracking logic that moves the calendar basically does nothing on
subsequent runs.