🐛 src: Fixes a formatted date bug in different time zone, e.g. PST
To ensure datetime consistency between the local browsers and the
datepicker, input date value can be a local datetime but it needs to
convert to UTC within the datepicker by picking just the date values
with `get*` methods instead of `getUTC*` methods.

FWIW, `get*` might return something unexpected in certain time zones
but that is acceptable since the datepicker is relying on what the
browser tells what it is.

Fixes tests to ensure all datetime values within the datepicker is
always in UTC time so `timeZone: UTC` is needed when doing a strict
comparison.

Also, added `_formatters` property to keep track of all kinds of date
formatters depending on value of `locale`.

Refactor year list with idiomatic JS and that requires polyfill for
newly added `Array.prototype.keys`.