✅ ci: Fixes tests due to SauceLabs' configuration on different browsers
SauceLabs runs certain browsers on a machine that returns a different
locale other than `en-US`. By defaulting to `en-US` for the tests should
resolve the issues.

Remove latest Node.js to reduce the number of testing environments as
SauceLabs timed out quite frequently as this might the cause of it.