✅ test: Adds more tests to perfect the datepicker
Fixes more bugs after adding more tests to perfecting the datepicker.

Fixes a local date bug when different timezones are taken into account.
Now there is a conversion for input date that reads local date reported
from the browser locally and UTC date is used internally. The datepicker
deals with UTC date to render the calendar and selected date is output
as UTC date as well in the format of `yyyy-MM-dd`. The fix has been
tested and it works so far in affected timezone (PST).

Upgrades `lit-element` to latest version - RC5 as of this commit.