🐛 src: Fixes `min` and `max` cannot be reset by setting falsy values
To match the behavior of native implementation of `input[type=date]` on
modern browsers, `min` and `max` can be reset by setting a falsy value.
However, this does not mean that `min` and `max` are implemented based
on the specs. The reset capability is important in this case thus it can
be implemented.

Also, updates tests for this new capability.