✅ src: Fixes a test issue where disabled days affect the outcome
One of the tests have been affected by an external factor that should
not be part of the test which is the `disabledDays`. The property comes
with a default value of `0,6`. Today happens to be one of the disabled
weekdays, renders the test to fail. Now, the property is reset before
running the affected test.

Adds a new helper function with overloading types to split string for
different kind of properties: `disabledDay` and `disabledDates`. Each
returns a different type after splitting and mapping with a callback.

Updates to use `static styles` class field.