♻️ src: Minor refactoring to expose disable days of each calendar
Minor refactoring on calendar method to expose a list of disable days
and cache the value on each computation before rendering. This helps in
determining if current focused date is one of the disabled days in
keyboard support method.

Keyboard support needs a rewrite to solve some of the issues around
`min`, `max`, and disabled days when arrow keys or other function keys
is used to navigate in the calendar.