✅ test: Adds moar tests for keyboard support.
Writing tests is awesome. The great thing about it is that you are given
chances to stress-test your code to an extent where it might break due
to some real world usage or inappropriate usage. It should fail when an
usage should be considered inappropriate by either throwing errors or
even halts the execution when necessary. The tests written here are
getting moar intense and moar comprehensive in some ways that it might
seem to be awfully daunting to everyone.

🐛 Fixes moar bugs to improve keyboard support of the calendar.