✅ test: Adds more tests for datepicker dialog.
Minor updates on datepicker dialog.

`?<property>` does not necessarily updates the value of the `<property>`
as it acts like setting an attribute. By assigning `<property>` a
`false` does not trigger updating the property that the change should
affect binded property to any child elements. `.<property>` works well
in this scenario.