💥 src: Year list updates according to changes in `min` and `max`
For every change in `min` and `max`, it will trigger updating
`_yearList`. Another notable improvement is that `min` and `max` now
come with default value when set with invalid date string:

  - `min` defaults to today's date.
  - `max` defaults to `2020-12-31`.

`min` and `max` now return empty string when it sets with invalid date
string such as null, '', and other non-date strings.