🐛 src: Fix blank calendar due to overloaded click handlers
When spamming the buttons to switch between calendar months, this queues
multiple task in the task queues. Each task queue is responsible for
updating the layout to determine whether or not buttons should be
rendered on screen for users to switch to the previous/ next calendar
month.

Some of the queuing tasks might be invalid if the new selected
date is less than the `_min` value or larger than the `_max` value. This
in turns causes an issue where animation continues to switch to the
subsequent calendar month and apparently that results in a blank
calendar.

To avoid such behavior, the patch adds a condition check to ensure that
the new selected date always falls between the defined value of `_min`
and `_max`. If the value happens to be out-of-range value, the
subsequent animation task(s) which is(are) already in the queue not get
executed.

Such fix avoids adding additional delay between the animations switching
between calendar months instead of implementing some kind of debouncing
or throttling approach. So, happy spamming the animations! 😄🎉