🐛 src: Criticl bug fixes due to invalid `_selectedDate`
On each navigation to next calendar month, the month value of the new
`_selectedDate` increments by one. This causes an issue where the day
value remains unchanged can lead to an non-existing date of the new
selected month, e.g. `2018-03-31` -> increment month -> `2018-04-31`.

Another bug fix is caused by newly added condition check to fix invalid
calendar being shown. The condition check only compares the months
between the `_min`, `_max` and new `_selectedDate` but it fails when
a calendar year comes into view which resets the calendar month which
could be smaller or bigger than the month of `_min` or `_max`.