🐛 src: Fix bugs that break the datepicker
Fix a bug where keyboard event that happens on datepicker body that
triggers on clickable navigate buttons when `Enter` or `Space` is
pressed. As a result, the updated focused date is no longer correct
due to unexpected update in the month value. The patch moves the focus
back to calendar view.

Fix a bug in `updated()` method where datepicker triggers updating the
position of calendar when `yearList` view is selected. Minor code
refactoring in `updated()` method to ensure private `updateCalendarPos`
method only runs

  - When `_startView` is `START_VIEW.CALENDAR`, or
  - When `landscape` changes.

Also, update tests due to the fixes above.