🐛 src: Calendars could be null if `min` > `value`
Minor improvement on `firstUpdated` and `updated` method to reduce
redundant workloads.

Besides minor cleanup, this also fixes an issue where an assumption was
made that the second calendar is unlikely to be null. However, this is
not true when `min` > `value`. Due to lack of strict checking on various
properties that depend on each other like `value`, `min`, and `max`,
strict checking increase the overall code complexity. As such, this
breaks when `disabledDaysSet` needs a fallback of empty Set when all
calendars are null.