🐛 src: Fix not updating focused date in other locale, fix #151
When a locale other than `en-US` is selected, the `day` property is no
longer a number in string but containing other character such as `24日`.
As such, it fails to be coerce to a number when updating focused date.

Unused `fullDate` property can be helpful in such case:

```
  // --snip

  /** Assign with new Date object with new `fullDate` value */
  this._focusedDate = new Date(selectedDayEl.fullDate);
```

New tests has been added to ensure the patch fixes the issue.